requirements.txt
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
